requirements.txt
asgiref==3.2.5
beautifulsoup4==4.8.2
boto3==1.9.96
botocore==1.12.96
certifi==2018.10.15
chardet==3.0.4
cycler==0.10.0
dj-database-url==0.5.0
Django==3.0.4
django-crispy-forms==1.7.2
django-environ==0.4.5
django-filter==2.2.0
django-heroku==0.3.1
django-ipware==2.1.0
django-storages==1.7.1
djangorestframework==3.11.0
docutils==0.14
gunicorn==19.9.0
idna==2.7
jmespath==0.9.3
joblib==0.14.1
kaggle==1.5.6
kiwisolver==1.1.0
lxml==4.5.0
Markdown==3.2.1
matplotlib==3.2.1
numpy==1.18.2
pandas==1.0.3
Pillow==5.2.0
psycopg2==2.7.7
pyparsing==2.4.6
python-dateutil==2.8.0
python-dotenv==0.12.0
python-slugify==4.0.0
pytz==2018.5
requests==2.19.1
s3transfer==0.2.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
selenium==3.141.0
six==1.12.0
sklearn==0.0
soupsieve==2.0
sqlparse==0.3.1
text-unidecode==1.3
tqdm==4.43.0
urllib3==1.23
whitenoise==4.1.2
